+++ app/src/main/java/org/glucosio/android/db/WeightReading.java
@@ -20,31 +20,31 @@
 
 package org.glucosio.android.db;
 
-import java.util.Date;
-
 import io.realm.RealmObject;
 import io.realm.annotations.PrimaryKey;
+
+import java.util.Date;
 
 public class WeightReading extends RealmObject {
     @PrimaryKey
     private long id;
 
-    private int reading;
+    private double reading;
     private Date created;
 
     public WeightReading() {
     }
 
-    public WeightReading(int reading, Date created) {
+    public WeightReading(double reading, Date created) {
         this.reading = reading;
         this.created = created;
     }
 
-    public int getReading() {
+    public double getReading() {
         return reading;
     }
 
-    public void setReading(int reading) {
+    public void setReading(double reading) {
         this.reading = reading;
     }
