+++ app/src/main/java/org/glucosio/android/service/DataLayerListenerService.java
@@ -25,11 +25,9 @@ import android.app.PendingIntent;
 import android.content.Intent;
 import android.support.v4.app.NotificationCompat;
 import android.util.Log;
-
 import com.google.android.gms.wearable.MessageEvent;
 import com.google.android.gms.wearable.WearableListenerService;
 import com.google.firebase.crash.FirebaseCrash;
-
 import org.glucosio.android.R;
 import org.glucosio.android.activity.MainActivity;
 import org.glucosio.android.db.DatabaseHandler;
@@ -65,7 +63,7 @@ public class DataLayerListenerService extends WearableListenerService {
     }
 
     private void addToDatabase(String reading, String readingType) {
-        int glucoseValue = Integer.parseInt(reading);
+        double glucoseValue = Double.parseDouble(reading);
         Calendar cal = Calendar.getInstance();
         DatabaseHandler dB = new DatabaseHandler(this);
         GlucoseReading gReading = new GlucoseReading(glucoseValue, readingType, cal.getTime(), "");
