+++ app/src/main/java/org/glucosio/android/tools/InputFilterMinMax.java
@@ -27,22 +27,22 @@ import android.util.Log;
 public class InputFilterMinMax implements InputFilter {
     private static final String TAG = "InputFilterMinMax";
 
-    private int min, max;
+    private double min, max;
 
-    public InputFilterMinMax(int min, int max) {
+    public InputFilterMinMax(double min, double max) {
         this.min = min;
         this.max = max;
     }
 
     public InputFilterMinMax(String min, String max) {
-        this.min = Integer.parseInt(min);
-        this.max = Integer.parseInt(max);
+        this.min = Double.parseDouble(min);
+        this.max = Double.parseDouble(max);
     }
 
     @Override
     public CharSequence filter(CharSequence source, int start, int end, Spanned dest, int dstart, int dend) {
         try {
-            int input = Integer.parseInt(dest.toString() + source.toString());
+            double input = Double.parseDouble(dest.toString() + source.toString());
             if (isInRange(min, max, input))
                 return null;
         } catch (NumberFormatException nfe) {
@@ -51,7 +51,7 @@ public class InputFilterMinMax implements InputFilter {
         return "";
     }
 
-    protected boolean isInRange(int a, int b, int c) {
+    protected boolean isInRange(double a, double b, double c) {
         return b > a ? c >= a && c <= b : c >= b && c <= a;
     }
 }
